add validator utils
